bash-completion: Backport ssh_config Include directive parsing
openssh 7.3p1 and up have the ability to specify additional config files
in an Include directive. This is supported on the stock openssh
implementation in macOS 10.14.6 at least.

Version 2.5 and onward of bash-completion have the ability to parse these
Include directives for hostname completion but require an upgraded version
of bash (4 or 5 specifically) to utilize on any version of macOS.

That said, nothing about the implementation of Include directive parsing
required bash 4 so this commit pulls down a patch file which applies it
back to the version of bash-completion provided by homebrew.